Изменил тип

+++ src/game.py
@@ -14,7 +14,7 @@ from pygame import Surface
 from pygame.time import Clock
 
 from src.state import State
-from src.modules import Pinger
+from src.pinger import Pinger
 
 from src.states import *  # Импорт НЕ УДАЛЯТЬ. Нужен чтобы все дочерние классы State были инициализированны
 
@@ -48,7 +48,7 @@ class Game:
         self.mouse_offset: tuple[int, int] = (0, 0)
 
         self.pinger: Pinger = Pinger()
-        self.last_ping_time: int = 0
+        self.last_ping_time: float = 0
 
         pg.display.set_caption('Город светофоров')
